ECS ClickableComponent
